To Submit

+++ src/main/java/com/bootcamp/project/service/UserService.java
@@ -120,6 +120,8 @@ public class UserService implements UserDetailsService {
     }
     public Boolean deleteUserByID(UUID userID){
         User user = findByUserID(userID);
+        if (user == null)
+            throw new ProjectException("User not Found");
         taskListService.deleteTasksLists(user);
         shoppingListService.deleteShoppingLists(user);
         deleteUser(user);
